renamed Prompt to Prompting
took it out of the model folder due to naming collision with the actual AR model
