+++ lib/langchainrb_rails.rb
@@ -6,8 +6,8 @@ require "rails"
 require_relative "langchainrb_rails/version"
 require "langchainrb_rails/railtie"
 require "langchainrb_rails/config"
+require 'langchainrb_rails/prompting'
 require_relative "langchainrb_overrides/vectorsearch/pgvector"
-require 'langchainrb_rails/models/prompt'
 
 module LangchainrbRails
   class Error < StandardError; end
