+++ lib/langchainrb_rails/generators/langchainrb_rails/templates/prompt_model.rb.tt
@@ -1,3 +1,3 @@
 class Prompt < ApplicationRecord
-  include LangchainrbRails::Models::Prompt
+  include LangchainrbRails::Prompting
 end
